+++ History.md
@@ -2,7 +2,7 @@
 Release date: unreleased
 
 * Dropped support for Ruby 2.7, 3.0+ is now required
-* Dropped support for Selenium < 4.1
+* Dropped support for Selenium < 4.8
 
 # Version 3.39.2
 Release date: 2023-06-10